🔧:refactor

+++ app/src/main/kotlin/jp/eijenson/connpass_searcher/view/presenter/MyJobServicePresenter.kt
@@ -29,7 +29,7 @@ class MyJobServicePresenter(
                     service.log("onError")
                 },
                 onNext = { result ->
-                    service.showNotification(result.id.toInt(), result.keyword, result.count.toInt())
+                    service.showNotification(result.id, result.keyword, result.count)
 
                 }
             )
